+++ src/Field/Definition/BloksField.php
@@ -186,4 +186,13 @@ final class BloksField extends AbstractField
 		$dataVisitor?->onDataVisit($this, $transformed);
 		return $transformed;
 	}
+
+	/**
+	 * @inheritDoc
+	 */
+	#[\Override]
+	public function useAsAdminDisplayName (?bool $canSync = null, ) : static
+	{
+		throw new InvalidFieldConfigurationException("This field type cannot be used as admin display name. Only text-based fields can be used.");
+	}
 }