+++ src/Field/Definition/BooleanField.php
@@ -5,6 +5,7 @@ namespace Torr\Storyblok\Field\Definition;
 use Symfony\Component\Validator\Constraints\NotNull;
 use Symfony\Component\Validator\Constraints\Type;
 use Torr\Storyblok\Context\ComponentContext;
+use Torr\Storyblok\Exception\InvalidFieldConfigurationException;
 use Torr\Storyblok\Field\FieldType;
 use Torr\Storyblok\Visitor\DataVisitorInterface;
 
@@ -66,5 +67,14 @@ final class BooleanField extends AbstractField
 
 		$dataVisitor?->onDataVisit($this, $data);
 		return $data;
+	}
+
+	/**
+	 * @inheritDoc
+	 */
+	#[\Override]
+	public function useAsAdminDisplayName (?bool $canSync = null, ) : static
+	{
+		throw new InvalidFieldConfigurationException("This field type cannot be used as admin display name. Only text-based fields can be used.");
 	}
 }